refactor: move interpolate() to utils/template, decouple prompts/commands/http/tools from workflows
The {{token}} interpolation helpers (interpolate/deepInterpolate) lived in
src/workflows/template.ts, so four non-workflow domains — prompts/resolver,
worker commands/runner, http/prompt-templates, and tools/prompt-templates —
depended on the workflow engine module just for a pure string helper.

Move the implementation to src/utils/template.ts (a leaf, zero-dep module that
worker code is already allowed to import) and repoint those four importers.
src/workflows/template.ts becomes a thin re-export so the engine (deepInterpolate)
and existing tests are untouched.

Cuts the prompts->workflows dependency edge (cycle-break #5 from the monorepo plan).
