+++ src/tools/prompt-templates/preview.ts
@@ -2,7 +2,7 @@ import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import * as z from "zod";
 import { getTemplateDefinition } from "@/prompts/registry";
 import { createToolRegistrar } from "@/tools/utils";
-import { interpolate } from "@/workflows/template";
+import { interpolate } from "@/utils/template";
 
 export const registerPreviewPromptTemplateTool = (server: McpServer) => {
   createToolRegistrar(server)(
